Removed old lima model
The old lima model was a placeholder that we want to remove
with some better data.
Camilo provided some better files that we can include.

The current sara file for Valparaiso is moved to a subfolder
to indicate the approach on how it was created.
This is the one that Raquel Zafrir did in her master thesis.